Move endpoint fixtures into correct conftest
