+++ back/test/endpoint_tests/conftest.py
@@ -1,5 +1,12 @@
 import pytest
 from auth import mock_user_for_request
+from boxtribute_server.app import create_app, register_blueprints
+from boxtribute_server.db import db
+from boxtribute_server.models import MODELS
+
+# It's crucial to import the blueprints from the routes module (NOT the blueprints
+# module) because only then they actually have routes registered.
+from boxtribute_server.routes import api_bp, app_bp, shared_bp
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -23,3 +30,34 @@ def auth_service(module_mocker):
 def unauthorized(mocker):
     """Effectively remove any permissions from current client."""
     mock_user_for_request(mocker, permissions=[])
+
+
+@pytest.fixture
+def app(setup_testing_database):
+    """Function fixture to create the Flask back-end on top of the given database
+    fixture."""
+    app = create_app()
+    app.debug = True
+    # Omit registering before-/teardown-request handlers (cf. main()). This is crucial
+    # to enable the usage of transaction rollbacks in the client fixture. Usually the
+    # handlers open/close the database connection but since this cannot happen during
+    # the transaction in the client fixture, they are rendered ineffective.
+    register_blueprints(app, api_bp, app_bp, shared_bp)
+    db.database = setup_testing_database
+    db.replica = None
+    with app.app_context():
+        yield app
+
+
+@pytest.fixture
+def client(app, setup_testing_database):
+    """Function fixture for any tests that include arbitrary operations on the database.
+    Use for testing GraphQL queries and mutations, and data model insertions/updates.
+    The fixture returns an app client that simulates sending requests to the app.
+    After each test, the applied database changes are rolled back.
+    The client's authentication and authorization may be separately defined or patched.
+    """
+    with setup_testing_database.bind_ctx(MODELS, False, False):
+        with setup_testing_database.atomic() as txn:
+            yield app.test_client()
+            txn.rollback()
